Preinitialization sequence added. Now windows is not necessary at all.
This preinit sequence sets the device to work correctly afterwards. The windows driver executes this sequence when the device is plugged in. This sequence can be repeated with no apparent negative effect.

The second received byte changes in a still unknown pattern, but it appears to not affect the behavior of the device in any way.

Still missing:

- Audio capture
- Stopping capture correctly
- Changing video modes (resolution?, input channel, etc)
- Only composite input is captured